implement delete swipe

+++ app/src/main/java/com/daggery/nots/home/view/HomeFragment.kt
@@ -2,6 +2,7 @@ package com.daggery.nots.home.view
 
 import android.content.DialogInterface
 import android.os.Bundle
+import android.util.Log
 import android.view.*
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
@@ -23,6 +24,7 @@ import com.daggery.nots.utils.NotsVibrator
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
+import com.google.android.material.R.style.ThemeOverlay_Material3_MaterialAlertDialog_Centered
 
 // TODO: Check if DatabaseOperation by Referring to Note UUID is Possible
 // TODO: Load list when splash screen is shown
@@ -141,7 +143,17 @@ class HomeFragmentUtils(
     }
 
     fun deleteNote(note: Note) {
-        fragment.viewModel.deleteNote(note)
+        MaterialAlertDialogBuilder(fragment.requireContext(), ThemeOverlay_Material3_MaterialAlertDialog_Centered)
+            .setIcon(R.drawable.ic_delete)
+            .setTitle("Delete Note")
+            .setMessage("Are you sure want to delete this note?\nThis action cannot be undone.")
+            .setPositiveButton("Delete") { _, _ ->
+                fragment.viewModel.deleteNote(note)
+            }
+            .setNegativeButton("Cancel") { dialog, _ ->
+                dialog.dismiss()
+            }
+            .show()
     }
 
     // Conditionally display empty illustration and notes list
